+++ tests/unit/logger-unit-tests.ts
@@ -1,15 +1,15 @@
-const expect = require('chai').expect;
-const sinon  = require('sinon');
-const util   = require('util');
+import { describe } from 'mocha';
+import { expect } from 'chai';
+import sinon from 'sinon';
+import { inspect } from 'util';
+import { getLogger, LogLevel } from '../../src';
 
 /* eslint-disable no-console */
 describe('Index unit tests', () => {
-    let logger;
     let sandbox;
 
     before(() => {
         sandbox = sinon.createSandbox();
-        logger  = require('../../src/index');
     });
 
     beforeEach(() => {
@@ -21,37 +21,37 @@ describe('Index unit tests', () => {
     });
 
     describe('logger', () => {
-        let expectedCount = 11;
+        let expectedCount = 8;
         it('levels - trace', (done) => {
-            testLogLevel('trace', expectedCount, done);
+            testLogLevel(LogLevel.TRACE, expectedCount, done);
         });
         it('levels - debug', (done) => {
-            testLogLevel('debug', --expectedCount, done);
+            testLogLevel(LogLevel.DEBUG, --expectedCount, done);
         });
         it('levels - info', (done) => {
-            testLogLevel('info', --expectedCount, done);
+            testLogLevel(LogLevel.INFO, --expectedCount, done);
         });
         it('levels - warning', (done) => {
-            testLogLevel('warning', --expectedCount, done);
+            testLogLevel(LogLevel.WARNING, --expectedCount, done);
         });
         it('levels - error', (done) => {
-            testLogLevel('error', expectedCount = expectedCount - 2, done);
+            testLogLevel(LogLevel.ERROR, --expectedCount, done);
         });
         it('levels - critical', (done) => {
-            testLogLevel('critical', expectedCount = expectedCount - 2, done);
+            testLogLevel(LogLevel.CRITICAL, --expectedCount, done);
         });
         it('levels - audit', (done) => {
-            testLogLevel('audit', expectedCount = expectedCount - 2, done);
+            testLogLevel(LogLevel.AUDIT, --expectedCount, done);
         });
         it('levels - audit_alert', (done) => {
-            testLogLevel('audit_alert', expectedCount, done);
+            testLogLevel(LogLevel.AUDIT_ALERT, --expectedCount, done);
         });
         it('levels - suppress', (done) => {
-            testLogLevel('suppress', expectedCount = expectedCount - 2, done);
+            testLogLevel(LogLevel.SUPPRESS, --expectedCount, done);
         });
 
         function testLogLevel(level, expectedCount, done) {
-            const log = logger({ level: level });
+            const log = getLogger({logLevel: level});
             Object.keys(log).forEach(key => {
                 if (typeof log[key] === 'function') {
                     log[key]('logging');
@@ -62,16 +62,16 @@ describe('Index unit tests', () => {
         }
 
         it('with attachment', (done) => {
-            const log        = logger();
-            const attachment = { foo: 'bar' };
+            const log = getLogger();
+            const attachment = {foo: 'bar'};
             log.info('test', attachment);
             expect(console.log.callCount).to.equal(1);
-            expect(console.log.firstCall.args[1]).to.equal(util.inspect(attachment));
+            expect(console.log.firstCall.args[1]).to.equal(inspect(attachment));
             done();
         });
 
         it('with timestamp', (done) => {
-            const log = logger({ timestamp: true });
+            const log = getLogger({timestamp: true});
             log.info('test');
             expect(console.log.callCount).to.equal(1);
             expect(console.log.firstCall.args[0]).to.match(/\[INFO] [-0-9]{10}T[:0-9]{8}.[0-9]{3}Z test/);
@@ -79,39 +79,37 @@ describe('Index unit tests', () => {
         });
 
         it('with prefix', (done) => {
-            const log = logger({ prefix: 'MyPrefix' });
+            const log = getLogger({prefix: 'MyPrefix'});
             log.info('test');
             expect(console.log.callCount).to.equal(1);
             expect(console.log.firstCall.args[0]).to.contain('MyPrefix');
             done();
         });
 
-        it('with invalid log level should default to info', (done) => {
-            const log = logger({ level: 'invalid' });
-            log.debug('test');
-            log.info('test');
-            expect(console.log.callCount).to.equal(1);
-            done();
-        });
-
         it('with log level via env', (done) => {
             process.env.LOG_LEVEL = 'ERROR';
-            const log             = logger();
-            log.info('test');
-            log.err('test');
-            expect(console.log.callCount).to.equal(1);
+            const log = getLogger();
+            expect(log.settings.logLevel).equals(LogLevel.ERROR);
             done();
         });
 
-        it('is debug enabled', (done) => {
+        it('with log level via env - should default to INFO', (done) => {
+            process.env.LOG_LEVEL = 'FOO';
+            const log = getLogger();
+            expect(log.settings.logLevel).equals(LogLevel.INFO);
+            done();
+        });
+
+        it('isDebugEnabled - Enabled', (done) => {
             process.env.LOG_LEVEL = 'DEBUG';
-            const log             = logger();
+            const log = getLogger();
+            expect(log.settings.logLevel).equals(LogLevel.DEBUG);
             expect(log.isDebugEnabled()).to.equal(true);
             done();
         });
-        it('is debug disabled', (done) => {
+        it('isDebugEnabled - Disabled', (done) => {
             process.env.LOG_LEVEL = 'INFO';
-            const log             = logger();
+            const log = getLogger();
             expect(log.isDebugEnabled()).to.equal(false);
             done();
         });